backend/requirements.txt
# Core Framework
fastapi
uvicorn
python-dotenv

# Transformers & Keras Dependencies
tf-keras
transformers
tensorflow

# Machine Learning & Data Science
scikit-learn
numpy

# LangChain & RAG Components
langchain
langchain-core
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers

# API & Utilities
pydantic
python-multipart
huggingface-hub
requests
python-jose
typing-extensions